# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  displayName?: string;
  sessionToken?: string;
  error?: string;
}

export interface GradeItem {
  courseCode: string;
  courseName: string;
  credits: string;
  grade: string;
  midtermGrade: string;
  points: string;
  section: string;
  subType: string;
}

export interface GradesResponse {
  grades: GradeItem[];
  fetchedAt: string;
  displayName?: string;
}

export interface ErrorResponse {
  error: string;
  needsReAuth?: boolean;
}

export interface SuccessResponse {
  success: boolean;
}