# lib/api-zod/src/generated/types/gradeItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GradeItem {
  courseCode: string;
  courseName: string;
  credits: string;
  grade: string;
  midtermGrade: string;
  points: string;
  section: string;
  subType: string;
}