# lib/api-zod/src/generated/types/gradesResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GradeItem } from './gradeItem';

export interface GradesResponse {
  grades: GradeItem[];
  fetchedAt: string;
  displayName?: string;
}